requirements.txt
PyQt6==6.10.2
PyQt6-sip==13.11.0
PyQt6-Qt6==6.10.1
Pillow==12.1.0
